RESTRUCTURED KNOWLEDGE DATATYPE; INFERENCE BUG FIX
So the story is a bit long. When implementing the bidirectional
inference function, it slowly dawned on me that the whole structure of
the `Knowledge` and `State` type is wrong. When the monad (State a) is
in a panic state, it should no longer be carrying the object of type a,
but should act like `Maybe`. Therefore I restructured the whole file.

After that, I discovered that the bidirectional inference process should
use the `unify` function instead of `match`. And, the unified
meta-variables didn't get replaced, resulting in an incorrect
derivation, where the meta-variable instantiation is all wrong. This
error will cause the derivation checker to reject the generated
derivation, since the meta-variable assignment context is erased after
the inference process.

In investigation of the above error, I discovered that the `mergeAssoc`
function does not produce idempotent substitutions. For instance,
  { x -> S y, z -> Z } and { z -> Z, y -> Z }
merges to
  { x -> S y, z -> Z, y -> Z }.
However, it is desirable to have (x -> S Z), i.e. to have y substituted
instead. This is now fixed.