WIP: IMPLEMENT BACKTRACKING
There is a hole in the program, of the type

> (Monad s) => (s j -> [s d]) -> (s [j] -> [s [d]])

I have no idea how I can implement that just yet. But I believe this
does not concern the nature of the type variables. It is a sort of
'blending lift' operation with a monad and a list (backtracking monad)
involved.

Basically, this function applies the given function (f :: s j -> [s d])
to each of the j's under the s monad, in that order, which is like mapM.
But we need to keep track of the case splits here..